Various error handling completed

+++ Tools/Ingestor.py
@@ -38,7 +38,8 @@ class Ingestor(IngestorInterface):
         #data = open('Tests/' + fileName, 'rb').read()
         data = open(fileName, 'rb').read()
         response = requests.put('https://platform.adobe.io/data/foundation/import/batches/' + batchId + '/datasets/' + datasetId + '/files/' + os.path.basename(fileName), headers=headers, data=data)
-        print(response)
+        if not self.error_check(response):
+            return
         #Signals the completion of the batch
         headers = {
             'x-gw-ims-org-id': imsOrg,
@@ -50,17 +51,20 @@ class Ingestor(IngestorInterface):
         )
         print('Signal Completion: ')
         response = requests.post('https://platform.adobe.io/data/foundation/import/batches/' + batchId, headers=headers, params=params)
-        print(response)
+        if not self.error_check(response):
+            print("Signal Completion has failed for " + fileName)
+        else:
+            print(fileName + " upload completed successfully")
         cataloguer.report(batchId, imsOrg, accessToken, apiKey)
 
-    def uploadLarge(self, fileName, dataSetId, imsOrg, accessToken:AuthToken, apiKey, cataloguer):
+    def uploadLarge(self, fileName, datasetId, imsOrg, accessToken:AuthToken, apiKey, cataloguer):
         headers = {
             'Content-Type': 'application/json',
             'x-gw-ims-org-id': imsOrg,
             'Authorization': 'Bearer ' + accessToken.getToken(),
             'x-api-key': apiKey
         }
-        data = '{ \n          "datasetId": "' + dataSetId + '" \n      }'
+        data = '{ \n          "datasetId": "' + datasetId + '" \n      }'
         response = requests.post('https://platform.adobe.io/data/foundation/import/batches', headers=headers, data=data)
         print('Create batch status: ' + response.json()['status'])
         batchId = response.json()['id']
@@ -77,9 +81,12 @@ class Ingestor(IngestorInterface):
             print('File upload of ' + os.path.basename(entry.path) + ' in progress')
             data = open(entry.path, 'rb').read()
             response = requests.put(
-                'https://platform.adobe.io/data/foundation/import/batches/' + batchId + '/datasets/' + dataSetId + '/files/' + os.path.basename(entry.path),
+                'https://platform.adobe.io/data/foundation/import/batches/' + batchId + '/datasets/' + datasetId + '/files/' + os.path.basename(entry.path),
                 headers=headers, data=data)
             print(response)
+            if not self.error_check(response):
+                print(os.path.basename(entry.path) + ' failed to upload')
+                continue
             os.remove(entry.path)
 
         headers = {
@@ -93,5 +100,14 @@ class Ingestor(IngestorInterface):
         print('Signal Completion: ')
         response = requests.post('https://platform.adobe.io/data/foundation/import/batches/' + batchId,
                                  headers=headers, params=params)
-        print(response)
+        if not self.error_check(response):
+            print("Signal Completion has failed for " + fileName)
+        else:
+            print(fileName + " upload completed Successfully")
         cataloguer.report(batchId, imsOrg, accessToken, apiKey)
+
+    def error_check(self, response):
+        if response.status_code != 200:
+            print("Error: " + response.status_code)
+            return False
+        return True
